Fix drawing of MMR general regions in JBIG2 images
CJBig2_Context::ParseGenericRegion() calls
page_->ComposeFromWithRect() with grd_->GetReplaceRect() as rect,
but for MMR images, GetReplaceRect() used to always return an
empty rect. That meant we decoded the image but then threw it away
when compositing it to the JBIG2 page.

To fix, set replace_rect_ to the whole bitmap at the end of MMR
decoding, similar to what ProgressiveDecodeArith() does.

Also add a few spec comments to FaxG4GetRow() (which I was reading while
debugging this.)

The test is a hand-draw bitmap that I converted to jbig2 format using
`jbig2` from the zip at https://www.itu.int/rec/T-REC-T.88-201808-I

I ran:

    jbig2 -i bitmap -f bmp \
          -o jbig2_mmr -F jb2 -ini jbig2-mmr.ini

with this jbig2-mmr.ini:

    -Gen -Seg 1
    -Gen -Param -Huff 1

I then ran the script linked to in the bug to convert it to a
PDF, and then manually edited it to convert it to .in format.
(Among other things, I ran the file through `xxd` and converted
the stream to /ASCIIHexDecode. I compared dumps from locally-added
hexdumping code and it looks like I didn't mess up the data.)

Bug: 432742364
Change-Id: I9df6c2c233ec89c439d0c813d9b67f64c36cd951
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/134270
Commit-Queue: Nico Weber <thakis@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Nico Weber <thakis@google.com>