Improve formfill_embeddertest failure messages
Because of the way the helpers were written, all test failures just
showed the line of the EXPECT that failed in the helper function, not
the line in the test that failed. So if you get a test failure you have
to edit the code to figure out what happened (potentially using
SCOPED_TRACE).

Instead the helper functions can just return the values they were
checking in the and use the EXPECT macros directly in the tests so that
the error messages have the correct line number.

Initially I tried out using testing::AssertionResult or making a custom
matcher. The former was too much rewriting and custom error messages
needed; the later is too awkward given the member variables needed.

Change-Id: I3d7985c7a460f873c9242a003ee79ff7ca4ece7e
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/135770
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Auto-Submit: April Kallmeyer <ask@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>